pyproject.toml
[tool.poetry]
name = "roborisk"
version = "0.1.0"
description = "Roborisk data ingestion"
authors = ["Roborisk Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
polygon-api-client = "*"
pydantic = "*"
sqlite-utils = "^3.38"
pydantic-settings = "*"
python-dotenv = "*"
numpy = "*"
numba = "*"
requests = "*"
rich = "*"
typer = {version = "*", extras = ["all"]}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
roborisk = "cli:main"
